backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.40.0
openai==1.58.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.3
Pillow==10.2.0
rembg==2.0.50
boto3==1.34.0
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
